Formula one PATCH Driver number service and persistence interface #510

+++ src/main/java/es/upm/miw/apaw_practice/adapters/rest/formula_one/DriverResource.java
@@ -20,7 +20,7 @@ public class DriverResource {
     }
 
     @PatchMapping(NUMBER_ID)
-    public Driver updatePoints(@PathVariable Integer number, @RequestBody Integer points) {
+    public Driver updatePoints(@PathVariable Integer number, @RequestBody Float points) {
         return this.driverService.updatePoints(number, points);
     }
 }
